fixed most tests one still has side effects - will fix later
