+++ src/app/app.component.xspec.ts
@@ -1,0 +1,42 @@
+/* tslint:disable:no-unused-variable */
+import { TestBed, async } from '@angular/core/testing';
+import { AppComponent } from './app.component';
+import { NavigationComponent } from './navigation/navigation.component';
+
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { APP_BASE_HREF } from '@angular/common';
+import { USERSERVICE, MockUserService, UserService } from './services';
+
+const routes: Routes = [
+];
+
+@NgModule({
+  imports: [
+    RouterModule.forRoot(routes)
+  ],
+  exports: [RouterModule]
+})
+export class AppRoutingModule { }
+
+describe('App: Web', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      declarations: [
+        AppComponent,
+        NavigationComponent
+      ],
+      imports: [AppRoutingModule],
+      providers: [
+        { provide: USERSERVICE, useClass: MockUserService },
+        { provide: APP_BASE_HREF, useValue: '/' }
+      ]
+    });
+  });
+
+  it('should create the app', async(() => {
+    let fixture = TestBed.createComponent(AppComponent);
+    let app = fixture.debugElement.componentInstance;
+    expect(app).toBeTruthy();
+  }));
+});
