+++ src/app/callback/callback.component.spec.ts
@@ -1,9 +1,25 @@
 /* tslint:disable:no-unused-variable */
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { DebugElement } from '@angular/core';
+import { NgModule, DebugElement } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { APP_BASE_HREF } from '@angular/common';
+import { USERSERVICE, MockUserService, UserService } from '../services';
 
 import { CallbackComponent } from './callback.component';
+
+
+const routes: Routes = [
+];
+
+@NgModule({
+  imports: [
+    RouterModule.forRoot(routes)
+  ],
+  exports: [RouterModule]
+})
+export class AppRoutingModule { }
+
 
 describe('CallbackComponent', () => {
   let component: CallbackComponent;
@@ -11,9 +27,14 @@ describe('CallbackComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ CallbackComponent ]
+      declarations: [CallbackComponent],
+      imports: [ AppRoutingModule ],
+      providers: [
+        { provide: USERSERVICE, useClass: MockUserService },
+        {provide: APP_BASE_HREF, useValue: '/'}
+      ]
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {
