+++ src/app/home/home.component.spec.ts
@@ -3,12 +3,6 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { DebugElement } from '@angular/core';
 
-
-import { Ng2UploaderModule } from 'ng2-uploader'
-import { MapIteratorPipe } from '../shared';
-import {APP_BASE_HREF} from '@angular/common';
-import { AppRoutingModule } from '../app.routing.module'
-
 import { HomeComponent } from './home.component';
 
 describe('HomeComponent', () => {
@@ -17,11 +11,7 @@ describe('HomeComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ HomeComponent, MapIteratorPipe ],
-      imports: [AppRoutingModule, Ng2UploaderModule],
-      providers: [
-        {provide: APP_BASE_HREF, useValue: '/'}
-      ]
+      declarations: [ HomeComponent ]
     })
     .compileComponents();
   }));
