+++ src/app/navigation/navigation.component.spec.ts
@@ -1,9 +1,23 @@
 /* tslint:disable:no-unused-variable */
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { DebugElement } from '@angular/core';
+import { NgModule, DebugElement } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { APP_BASE_HREF } from '@angular/common';
+import { USERSERVICE, MockUserService, UserService } from '../services';
 
 import { NavigationComponent } from './navigation.component';
+
+const routes: Routes = [
+];
+
+@NgModule({
+  imports: [
+    RouterModule.forRoot(routes)
+  ],
+  exports: [RouterModule]
+})
+export class AppRoutingModule { }
 
 describe('NavigationComponent', () => {
   let component: NavigationComponent;
@@ -11,9 +25,14 @@ describe('NavigationComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ NavigationComponent ]
+      declarations: [NavigationComponent],
+      imports: [AppRoutingModule],
+      providers: [
+        { provide: USERSERVICE, useClass: MockUserService },
+        { provide: APP_BASE_HREF, useValue: '/' }
+      ]
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {
